Rename to NativeWorkerAdapter
