+++ pyproject.toml
@@ -51,7 +51,7 @@ scaler_symphony_cluster = "scaler.entry_points.symphony_cluster:main"
 scaler_top = "scaler.entry_points.top:main"
 scaler_ui = "scaler.entry_points.webui:main"
 scaler_object_storage_server = "scaler.entry_points.object_storage_server:main"
-scaler_vm_worker_adapter = "scaler.entry_points.vm_adapter:main"
+scaler_native_worker_adapter = "scaler.entry_points.native_adapter:main"
 
 [project.optional-dependencies]
 uvloop = ["uvloop"]
